# back-end/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Noroff EP e-commerce API",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module."
  },
  "host": "localhost:3001",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/users/": {
      "get": {
        "tags": [
          "Users"
        ],
        "description": "Retrieve all users. Admin access is required.",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "List of users retrieved successfully.",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "success"
                },
                "statuscode": {
                  "type": "number",
                  "example": 200
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "result": {
                      "type": "string",
                      "example": "Users retrieved successfully."
                    },
                    "users": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "Id": {
                            "type": "number",
                            "example": 2
                          },
                          "FirstName": {
                            "type": "string",
                            "example": "John"
                          },
                          "LastName": {
                            "type": "string",
                            "example": "Doe"
                          },
                          "Username": {
                            "type": "string",
                            "example": "user123"
                          },
                          "Email": {
                            "type": "string",
                            "example": "johndoe@email.com"
                          },
                          "Address": {
                            "type": "string",
                            "example": "123 Street, City"
                          },
                          "Phone": {
                            "type": "string",
                            "example": "123456789"
                          },
                          "createdAt": {
                            "type": "string",
                            "example": "2024-12-21T14:29:40.000Z"
                          },
                          "updatedAt": {
                            "type": "string",
                            "example": "2024-12-21T14:29:40.000Z"
                          },
                          "MembershipId": {
                            "type": "number",
                            "example": 1
                          },
                          "RoleId": {
                            "type": "number",
                            "example": 2
                          },
                          "Membership": {
                            "type": "object",
                            "properties": {
                              "Id": {
                                "type": "number",
                                "example": 1
                              },
                              "Name": {
                                "type": "string",
                                "example": "Bronze"
                              },
                              "FromTotalPurchases": {
                                "type": "number",
                                "example": 0
                              },
                              "ToTotalPurchases": {
                                "type": "number",
                                "example": 15
                              },
                              "Discount": {
                                "type": "number",
                                "example": 0
                              }
                            }
                          },
                          "Cart": {},
                          "Orders": {
                            "type": "array",
                            "example": [],
                            "items": {}
                          },
                          "Role": {
                            "type": "object",
                            "properties": {
                              "Id": {
                                "type": "number",
                                "example": 2
                              },
                              "Name": {
                                "type": "string",
                                "example": "User"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal server error.",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "error"
                },
                "statuscode": {
                  "type": "number",
                  "example": 500
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "result": {
                      "type": "string",
                      "example": "Failed to retrieve users."
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/users/{id}": {
      "put": {
        "tags": [
          "Users"
        ],
        "description": "Update user details by ID. Admin access is required.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID of the user to be updated."
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Fields to update for the user.",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "roleId": {
                  "type": "number",
                  "example": 1
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User updated successfully.",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "success"
                },
                "statuscode": {
                  "type": "number",
                  "example": 200
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "result": {
                      "type": "string",
                      "example": "User updated successfully."
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "Invalid or missing fields.",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "error"
                },
                "statuscode": {
                  "type": "number",
                  "example": 400
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "result": {
                      "type": "string",
                      "example": "Required fields are missing or invalid."
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal server error.",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "error"
                },
                "statuscode": {
                  "type": "number",
                  "example": 500
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "result": {
                      "type": "string",
                      "example": "Failed to update user."
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
...
  },
...
}